+++ fitnex_frontend/app/page.tsx
@@ -1,31 +1,31 @@
 "use client"
 
-import Exercises from '@/components/Exercises'
+// import Exercises from '@/components/Exercises'
 import HeroBanner from '@/components/HeroBanner'
 import MaxWidthWrapper from '@/components/MaxWidthWrapper'
-import SearchExcercise from '@/components/SearchExcercise'
 import { BodyPartProps, ExerciseProps } from '@/lib/types'
-import Image from 'next/image'
+
+
 import { useState } from 'react'
+import SearchExcercise from '@/components/SearchExcercise'
+import HeroSection from '@/components/HeroSection'
+import OurFeatures from '@/components/OurFeatures'
+import Customizable from '@/components/Customizable'
+import MadeSimple from '@/components/MadeSimple'
+import MadeFor from '@/components/MadeFor'
+import Faqs from '@/components/Faqs'
+import NeedHelp from '@/components/NeedHelp'
+
 
 export default function Home() {
 
-  //changes are reflected acrross the app
-  const [exercises, setExercises] = useState<ExerciseProps[]>([])
-  const [bodyPart, setBodyPart] = useState<BodyPartProps[]>([]);
-
-  // console.log(bodyPart)
-
+  
   return (
     <>
       <MaxWidthWrapper>
         <HeroBanner />
 
-        <SearchExcercise setExercises={setExercises} bodyPart={bodyPart} setBodyPart={setBodyPart} />
 
-        <Exercises exercises={exercises} setExercises={setExercises} bodyPart={bodyPart}  />
-      </MaxWidthWrapper>
-      <div>
         <HeroSection />
         <OurFeatures />
         <Customizable />
@@ -33,7 +33,9 @@ export default function Home() {
         <MadeFor />
         <Faqs />
         <NeedHelp />
-      </div>
+      </MaxWidthWrapper>
+
+
     </>
   )
 }
